+++ packages/nextjs/components/safe-area-container.tsx
@@ -1,0 +1,23 @@
+import { SafeAreaInsets } from "@/types";
+
+interface SafeAreaContainerProps {
+  children: React.ReactNode;
+  insets?: SafeAreaInsets;
+}
+
+export const SafeAreaContainer = ({
+  children,
+  insets,
+}: SafeAreaContainerProps) => (
+  <main
+    className="flex min-h-screen flex-col items-center justify-center gap-y-3"
+    style={{
+      marginTop: insets?.top ?? 0,
+      marginBottom: insets?.bottom ?? 0,
+      marginLeft: insets?.left ?? 0,
+      marginRight: insets?.right ?? 0,
+    }}
+  >
+    {children}
+  </main>
+);
