+++ packages/nextjs/package.json
@@ -14,6 +14,10 @@
     "vercel:yolo": "vercel --build-env YARN_ENABLE_IMMUTABLE_INSTALLS=false --build-env ENABLE_EXPERIMENTAL_COREPACK=1 --build-env NEXT_PUBLIC_IGNORE_BUILD_ERROR=true --build-env VERCEL_TELEMETRY_DISABLED=1"
   },
   "dependencies": {
+    "@farcaster/auth-client": "^0.3.0",
+    "@farcaster/frame-node": "^0.0.22",
+    "@farcaster/frame-sdk": "^0.0.37",
+    "@farcaster/frame-wagmi-connector": "^0.0.25",
     "@heroicons/react": "~2.1.5",
     "@rainbow-me/rainbowkit": "2.1.6",
     "@tanstack/react-query": "~5.59.15",
