+++ packages/nextjs/types/index.ts
@@ -1,0 +1,6 @@
+export interface SafeAreaInsets {
+  top?: number;
+  bottom?: number;
+  left?: number;
+  right?: number;
+}
